+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Release 2024-09-10
+- Fix tests, add ridiculously simple `fillHourlyTimeSeriesWithNull` function 
+  just so my pea sized brain doesn't have to get it wrong again.  
+
 ## Release 2024-02-14
 - Added a test and example for `splitByIndex`.